chore(kuma-cp): move attachments to inspect package
