+++ pkg/api-server/inspect_endpoints.go
@@ -280,7 +280,7 @@ func inspectPolicies(
 				return
 			}
 			if matchedPolicies != nil {
-				for policy, attachments := range core_xds.GroupByPolicy(matchedPolicies, dp.Spec.Networking) {
+				for policy, attachments := range inspect.GroupByPolicy(matchedPolicies, dp.Spec.Networking) {
 					if policy.Type == resType && policy.Key.Name == policyName && policy.Key.Mesh == meshName {
 						attachmentList := []api_server_types.AttachmentEntry{}
 						for _, attachment := range attachments {
@@ -326,15 +326,15 @@ func routeDestinationToAPIDestination(des route.Destination) api_server_types.De
 }
 
 func newDataplaneInspectResponse(matchedPolicies *core_xds.MatchedPolicies, dp *core_mesh.DataplaneResource) []*api_server_types.DataplaneInspectEntry {
-	attachmentMap := core_xds.GroupByAttachment(matchedPolicies, dp.Spec.Networking)
+	attachmentMap := inspect.GroupByAttachment(matchedPolicies, dp.Spec.Networking)
 
 	entries := make([]*api_server_types.DataplaneInspectEntry, 0, len(attachmentMap))
-	attachments := []core_xds.Attachment{}
+	attachments := []inspect.Attachment{}
 	for attachment := range attachmentMap {
 		attachments = append(attachments, attachment)
 	}
 
-	sort.Stable(core_xds.AttachmentList(attachments))
+	sort.Stable(inspect.AttachmentList(attachments))
 
 	for _, attachment := range attachments {
 		entry := &api_server_types.DataplaneInspectEntry{
@@ -438,11 +438,11 @@ func newGatewayDataplaneInspectResponse(
 }
 
 func routeToPolicyInspect(
-	policyMap map[core_xds.PolicyKey][]envoy.Tags,
+	policyMap map[inspect.PolicyKey][]envoy.Tags,
 	des route.Destination,
-) map[core_xds.PolicyKey][]envoy.Tags {
+) map[inspect.PolicyKey][]envoy.Tags {
 	for kind, p := range des.Policies {
-		policyKey := core_xds.PolicyKey{
+		policyKey := inspect.PolicyKey{
 			Type: kind,
 			Key:  core_model.MetaToResourceKey(p.GetMeta()),
 		}
@@ -458,8 +458,8 @@ func routeToPolicyInspect(
 func gatewayEntriesByPolicy(
 	proxy core_xds.Proxy,
 	listenerInfos []gateway.GatewayListenerInfo,
-) map[core_xds.PolicyKey][]api_server_types.PolicyInspectEntry {
-	policyMap := map[core_xds.PolicyKey][]api_server_types.PolicyInspectEntry{}
+) map[inspect.PolicyKey][]api_server_types.PolicyInspectEntry {
+	policyMap := map[inspect.PolicyKey][]api_server_types.PolicyInspectEntry{}
 
 	dpKey := core_model.MetaToResourceKey(proxy.Dataplane.GetMeta())
 	resourceKey := api_server_types.ResourceKeyEntry{
@@ -467,13 +467,13 @@ func gatewayEntriesByPolicy(
 		Name: dpKey.Name,
 	}
 
-	listenersMap := map[core_xds.PolicyKey][]api_server_types.PolicyInspectGatewayListenerEntry{}
+	listenersMap := map[inspect.PolicyKey][]api_server_types.PolicyInspectGatewayListenerEntry{}
 	for _, info := range listenerInfos {
-		hostMap := map[core_xds.PolicyKey][]api_server_types.PolicyInspectGatewayHostEntry{}
+		hostMap := map[inspect.PolicyKey][]api_server_types.PolicyInspectGatewayHostEntry{}
 		for _, info := range info.HostInfos {
-			routeMap := map[core_xds.PolicyKey][]api_server_types.PolicyInspectGatewayRouteEntry{}
+			routeMap := map[inspect.PolicyKey][]api_server_types.PolicyInspectGatewayRouteEntry{}
 			for _, entry := range info.Entries {
-				entryMap := map[core_xds.PolicyKey][]envoy.Tags{}
+				entryMap := map[inspect.PolicyKey][]envoy.Tags{}
 				if entry.Mirror != nil {
 					entryMap = routeToPolicyInspect(entryMap, entry.Mirror.Forward)
 				}
@@ -533,7 +533,7 @@ func gatewayEntriesByPolicy(
 
 	if logging, ok := proxy.Policies.TrafficLogs[core_mesh.PassThroughService]; ok {
 		wholeGateway := api_server_types.NewPolicyInspectGatewayEntry(resourceKey, gatewayKey)
-		policyKey := core_xds.PolicyKey{
+		policyKey := inspect.PolicyKey{
 			Type: core_mesh.TrafficLogType,
 			Key:  core_model.MetaToResourceKey(logging.GetMeta()),
 		}
@@ -544,7 +544,7 @@ func gatewayEntriesByPolicy(
 	}
 	if trace := proxy.Policies.TrafficTrace; trace != nil {
 		wholeGateway := api_server_types.NewPolicyInspectGatewayEntry(resourceKey, gatewayKey)
-		policyKey := core_xds.PolicyKey{
+		policyKey := inspect.PolicyKey{
 			Type: core_mesh.TrafficTraceType,
 			Key:  core_model.MetaToResourceKey(trace.GetMeta()),
 		}
